requirements.txt
Flask==2.3.3
coremltools==7.1
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
pytesseract==0.3.10
psutil==5.9.6
transformers==4.35.0
torch==2.0.1
git+https://github.com/openai/CLIP.git
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
psycopg2-binary==2.9.7
supabase==2.3.4
python-dotenv==1.0.0
langchain-community
langchain-ollama
langchain-core 